A working function for logging in
It logs checks if username exists and if the password is correct. for
now it's case sensitive but it doesn't have to be. There's also a
function commented out for checking if user is an Admin or a Customer.
We can decide later if that is supposed to happen here in the log in or
somewhere else in the program.